dtype debug

+++ deep_field_metadetect/jaxify/tests/test_jax_metadetect.py
@@ -391,9 +391,7 @@ def test_metadetect_single_band_deep_field_metadetect_mfrac_deep():
         n_objs=10,
         return_dfmd_obs=True,
     )
-    obs_d = obs_d.replace(
-        mfrac=np.float32(rng.uniform(0.5, 0.7, size=obs_w.image.shape))
-    )
+    obs_d = obs_d.replace(mfrac=rng.uniform(0.5, 0.7, size=obs_w.image.shape))
 
     dk = compute_dk(image_size=nxy_psf, pixel_scale=scale)
     kim_size = compute_kim_size(image_size=nxy_psf)
